Attack of the drones!
Drones are a work in progress.  Due to the nature of the x y mapping
system, they move about the canvas rather than the map, but this will be
fixed soon.  Press space when near the castle to make a drone.  There is
a lot more I have to do to get them working right, but this is just so
you can see what I have so far.